Tooltip for textbox created for show range values. delete run and stop method inside drive service, write register method replaced that. create slider max value and mid value for binding, the max value will be updated with motor rated frequency and mid value is a dinamyc property. method for read drive status and digital input, now only read drive status. add validation values of max frequency, rated current, max speed and rated voltage, if that values equal to zero the run command can not executed. add validatios values for comport combobox, baudrate combobox and slave id textbox, if that values be empty or zero the connect command can not executed

+++ InvertekDrive-RTU-UI/InvertekDrive-RTU-UI/Model/MotorModel.cs
@@ -2,10 +2,10 @@
 
 public class MotorModel
 {
-    public double
+    public int
         RatedVoltage { get; set; }
     public double
         RatedCurrent { get; set; }
-    public double
+    public int
         RatedFrequency { get; set; }
 }
